configure: swap lib order in LIBMXFPP_LDADDLIBS
The wrapper script used for executables and generated by libtool
sets the LD_LIBRARY_PATH before running the executable. Prior to
this commit it would set the local library path (src/.libs) after
the /usr/local/lib. After this commit the local library path is set
first.

This was not an issue in libMXFpp but this change brings it inline with
bmx.